+++ test/Spec.hs
@@ -1,15 +1,17 @@
 {-# LANGUAGE BlockArguments #-}
 {-# LANGUAGE NumericUnderscores #-}
+{-# LANGUAGE PatternSynonyms #-}
 {-# LANGUAGE TemplateHaskell #-}
 
 import Hedgehog
 import qualified Hedgehog.Gen as Gen
 import qualified Hedgehog.Range as Range
 
-import Control.Monad (when)
+import Control.Monad (unless)
 import Control.Monad.IO.Class (MonadIO(..))
-import Data.Function (on)
-import Data.Bifunctor (bimap)
+import Data.Foldable (toList)
+import Data.Maybe (catMaybes)
+import GHC.Stack (withFrozenCallStack)
 import System.Exit (exitFailure, exitSuccess)
 import Text.Printf (printf)
 
@@ -17,36 +19,60 @@ import qualified Foreign
 import qualified Linear
 import qualified Geomancy
 import qualified Geomancy.Mat4
-import qualified Geomancy.Vulkan.Projection
 
-prop_multiply :: Property
-prop_multiply = withTests 10_000 $ property do
-  (nameA, a) <- forAll genTransform
-  annotate nameA
+import Linear ((!*!))
 
-  (nameB, b) <- forAll genTransform
-  annotate nameB
+main :: IO ()
+main = do
+  passed <- checkParallel discovered
+  if passed then
+    exitSuccess
+  else
+    exitFailure
 
-  let transMult = Geomancy.Mat4.transpose (a <> b)
-  annotateShow transMult
+-- | Enough tests to cover the principles
+pattern PROP_TESTS :: TestLimit
+pattern PROP_TESTS = 10_000
 
-  let multTrans = Geomancy.Mat4.transpose a <> Geomancy.Mat4.transpose b
-  annotateShow multTrans
+-- | Try harder to catch FP precision errors
+pattern PROP_TESTS_BRUTAL :: TestLimit
+pattern PROP_TESTS_BRUTAL = 10_000_000
+
+prop_assoc_multiply :: Property
+prop_assoc_multiply = withTests PROP_TESTS $ property do
+  (a, a_) <- forAllTransform
+  (b, b_) <- forAllTransform
+  (c, c_) <- forAllTransform
 
   let
-    delta =
-      Geomancy.Mat4.elementwise transMult multTrans \a b ->
-        abs (a - b)
-  annotateShow delta
+    ab'c = (a <> b) <> c
+    a'bc = a <> (b <> c)
+    delta' = nearlyEqualMat4 ab'c a'bc
+  annotateShow delta'
 
-  aLinear <- toM44 a
-  bLinear <- toM44 b
   let
-    mtLinear = Linear.transpose $ aLinear Linear.!*! bLinear
-    tmLinear = Linear.transpose aLinear Linear.!*! Linear.transpose bLinear
-    deltaLinear = abs $ mtLinear - tmLinear
-  delta' <- toM44 delta
-  delta' === deltaLinear
+    ab_c = (a_ !*! b_) !*! c_
+    a_bc = a_ !*! (b_ !*! c_)
+    delta_ = nearlyEqualM44 ab_c a_bc
+  annotateShow delta_
+
+  -- Intra-library transitivity
+  unless (null $ catMaybes delta') do
+    -- XXX: check only if there is some outstanding error
+    delta' === delta_
+
+  -- Inter-library calculated values nearlyEqual
+  ab'c_ <- toM44 ab'c
+  catMaybes (nearlyEqualM44 ab_c ab'c_) === []
+
+  a'bc_ <- toM44 a'bc
+  catMaybes (nearlyEqualM44 a_bc a'bc_) === []
+
+forAllTransform :: PropertyT IO (Geomancy.Mat4.Mat4, Linear.M44 Float)
+forAllTransform = withFrozenCallStack do
+  (_name, g) <- forAllWith fst genTransform
+  l <- toM44 g
+  pure (g, l)
 
 genTransform :: Gen ([Char], Geomancy.Mat4)
 genTransform = Gen.choice
@@ -88,14 +114,14 @@ genTransform = Gen.choice
         , Geomancy.Mat4.scale x y z
         )
 
-toVulkan :: Linear.M44 Float -> Linear.M44 Float
-toVulkan = {- Linear.transpose . -} (correction Linear.!*!)
-  where
-    correction = Linear.V4
-      (Linear.V4 1   0  0   0)
-      (Linear.V4 0 (-1) 0   0)
-      (Linear.V4 0   0  0.5 0.5)
-      (Linear.V4 0   0  0   1)
+-- toVulkan :: Linear.M44 Float -> Linear.M44 Float
+-- toVulkan = {- Linear.transpose . -} (correction Linear.!*!)
+--   where
+--     correction = Linear.V4
+--       (Linear.V4 1   0  0   0)
+--       (Linear.V4 0 (-1) 0   0)
+--       (Linear.V4 0   0  0.5 0.5)
+--       (Linear.V4 0   0  0   1)
 
 toM44 :: MonadIO io => Geomancy.Mat4 -> io (Linear.M44 Float)
 toM44 mat4 =
@@ -103,10 +129,20 @@ toM44 mat4 =
     Foreign.with mat4 $
       Foreign.peek . Foreign.castPtr
 
-main :: IO ()
-main = do
-  passed <- checkParallel $$(discover)
-  if passed then
-    exitSuccess
+type NearlyEqual = Maybe (Float, Float, Float)
+
+nearlyEqualMat4 :: Geomancy.Mat4.Mat4 -> Geomancy.Mat4.Mat4 -> [NearlyEqual]
+nearlyEqualMat4 a b = Geomancy.Mat4.zipWith nearlyEqual a b
+
+nearlyEqualM44 :: Linear.M44 Float -> Linear.M44 Float -> [NearlyEqual]
+nearlyEqualM44 a b = zipWith nearlyEqual (concatMap toList a) (concatMap toList b)
+
+nearlyEqual :: Float -> Float -> NearlyEqual
+nearlyEqual x y =
+  if x == y || abs (1 - x / y) < 0.001 then
+    Nothing
   else
-    exitFailure
+    Just (x, y, abs (1 - x / y))
+
+discovered :: Group
+discovered = $$(discover)